dynamic filters

+++ src/main/java/com/alkemy/challenge/disney/dto/ActorBasicDTO.java
@@ -3,12 +3,11 @@ package com.alkemy.challenge.disney.dto;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.util.List;
 
 @Getter
 @Setter
 public class ActorBasicDTO {
     private String name;
-    private Integer age;
-    private Integer weight;
+    private Long age;
+    private Long weight;
 }
