+++ src/main/java/com/alkemy/challenge/disney/entity/ActorEntity.java
@@ -24,8 +24,8 @@ public class ActorEntity {
     private Long id;
     private String image;
     private String name;
-    private Integer age;
-    private Integer weight;
+    private Long age;
+    private Long weight;
     private String story;
     private boolean deleted = Boolean.FALSE;
